add rating service and rename current rating service to rating_comparison
