+++ lib/lighthouse_bgs/services.rb
@@ -25,6 +25,7 @@ require 'lighthouse_bgs/services/manage_representative'
 require 'lighthouse_bgs/services/org'
 require 'lighthouse_bgs/services/person'
 require 'lighthouse_bgs/services/rating'
+require 'lighthouse_bgs/services/rating_comparison'
 require 'lighthouse_bgs/services/rating_profile'
 require 'lighthouse_bgs/services/standard_data'
 require 'lighthouse_bgs/services/vet_record'
